HW20200910开工，加入日志打印log4j
