--- src/src/xyz/huowang/hwblog/constants/ErrorConstant.java
@@ -1,12 +1,0 @@
-package src.xyz.huowang.hwblog.constants;
-
-/**
- * @author ：HW
- * @date ：Created in 2020/9/8 22:27
- * @description： 定义错误码，报错信息
- * @version: 1.0
- */
-public class ErrorConstant {
-
-    public static final String REGISTER_USER_ALREADY_EXIST = "注册的用户名已存在";
-}
