--- src/src/xyz/huowang/hwblog/constants/SqlConstant.java
@@ -1,10 +1,0 @@
-package src.xyz.huowang.hwblog.constants;
-
-/**
- * @author ：HW
- * @date ：Created in 2020/9/8 22:28
- * @description：TODO
- * @version: 1.0
- */
-public class SqlConstant {
-}
