--- src/src/xyz/huowang/hwblog/web/UI/RegisterUIServlet.java
@@ -1,25 +1,0 @@
-package src.xyz.huowang.hwblog.web.UI;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-/**
- * @author ：HW
- * @date ：Created in 2020/9/8 22:45
- * @description：TODO
- * @version: 1.0
- */
-public class RegisterUIServlet extends HttpServlet {
-
-    public void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        request.getRequestDispatcher("/WEB-INF/pages/Register.jsp").forward(request, response);
-    }
-    public void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        doGet(request, response);
-    }
-}
