+++ src/xyz/huowang/hwblog/constants/Constant.java
@@ -1,0 +1,12 @@
+package xyz.huowang.hwblog.constants;
+
+/**
+ * @author ：HW
+ * @date ：Created in 2020/9/8 22:26
+ * @description： java常量 Xxx.NAME
+ * @version: 1.0
+ */
+public class Constant {
+
+    public static final String ACCNO = "AccNo";
+}
