+++ src/xyz/huowang/hwblog/constants/SqlConstant.java
@@ -1,0 +1,10 @@
+package xyz.huowang.hwblog.constants;
+
+/**
+ * @author ：HW
+ * @date ：Created in 2020/9/8 22:28
+ * @description：TODO
+ * @version: 1.0
+ */
+public class SqlConstant {
+}
