+++ web/WEB-INF/pages/Login.jsp
@@ -12,7 +12,7 @@
     <title>用户登陆</title>
 </head>
 <body>
-<form action="${pageContext.request.contextPath }/servlet/LoginServlet" method="post">
+<form action="${pageContext.request.contextPath }/LoginServlet" method="post">
     用户名：<input type="text" name="username"><br/>
     密码：<input type="password" name="password"><br/>
     <input type="submit" value="登陆">
