+++ web/WEB-INF/web.xml
@@ -5,28 +5,48 @@
          version="4.0">
     <display-name>adminsystem1</display-name>
     <welcome-file-list>
-        <welcome-file>Welcome.html</welcome-file>
-        <welcome-file>Welcome.htm</welcome-file>
-        <welcome-file>Welcome.jsp</welcome-file>
+        <welcome-file>index.html</welcome-file>
+        <welcome-file>index.htm</welcome-file>
+        <welcome-file>index.jsp</welcome-file>
     </welcome-file-list>
 
     <servlet>
         <servlet-name>RegisterUIServlet</servlet-name>
-        <servlet-class>src.xyz.huowang.hwblog.web.UI.RegisterUIServlet</servlet-class>
+        <servlet-class>xyz.huowang.hwblog.web.UI.RegisterUIServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>RegisterUIServlet</servlet-name>
-        <url-pattern>/register</url-pattern>
+        <url-pattern>/RegisterUIServlet</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>LoginUIServlet</servlet-name>
-        <servlet-class>src.xyz.huowang.hwblog.web.UI.LoginUIServlet</servlet-class>
+        <servlet-class>xyz.huowang.hwblog.web.UI.LoginUIServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
         <servlet-name>LoginUIServlet</servlet-name>
-        <url-pattern>/login</url-pattern>
+        <url-pattern>/LoginUIServlet</url-pattern>
+    </servlet-mapping>
+
+    <servlet>
+        <servlet-name>RegisterServlet</servlet-name>
+        <servlet-class>xyz.huowang.hwblog.web.controller.RegisterServlet</servlet-class>
+    </servlet>
+
+    <servlet-mapping>
+        <servlet-name>RegisterServlet</servlet-name>
+        <url-pattern>/RegisterServlet</url-pattern>
+    </servlet-mapping>
+
+    <servlet>
+        <servlet-name>LoginServlet</servlet-name>
+        <servlet-class>xyz.huowang.hwblog.web.controller.LoginServlet</servlet-class>
+    </servlet>
+
+    <servlet-mapping>
+        <servlet-name>LoginServlet</servlet-name>
+        <url-pattern>/LoginServlet</url-pattern>
     </servlet-mapping>
 </web-app>
